requirements.txt
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
SQLAlchemy==2.0.36
